Add svg path class
This allows to store SVG paths and to calculate their bounding box.

This will be used for describing display cutouts as the bounding box
describes an upper bound of the occluded area.

We currently support a subset of the path spec omitting Q/q, S/s and T/t
which we'll add later on.