Add cutout class to store display panel cutouts
A cutout describes an area of a display panel that is not usable for
rendering. E.g. because there are other elements like cameras. The
cutout itself is described via a closed SVG path. This allows shells
compsitors to make decision on where to place windows and ui elements.
See the class description for more details.